Fixed rotation scaling issues by applying transformation "matrices" to the raw vertex data. Scaling MUST happen before rotation, otherwise the shape is deformed.

+++ src/main/java/org/example/Components/Transform.java
@@ -3,9 +3,9 @@ package org.example.Components;
 public class Transform {
 	public double x;
 	public double y;
-	public float angle;
+	public double angle;
 
-	public Transform(double x, double y, float angle) {
+	public Transform(double x, double y, double angle) {
 		this.x = x;
 		this.y = y;
 		this.angle = angle;
